refactoring: 훅 분리
