+++ src/components/cs/CsPage.tsx
@@ -1,22 +1,15 @@
 'use client'
 
-import { useMemo } from 'react'
 import CsDetailBox from './CsDetailBox'
 import SubmittedDetailBox from './SubmittedDetailBox'
-import { useCsProblemDetailQuery } from '@/api/cs'
+import { useCsPage } from '@/hooks/cs/useCsPage'
 
 interface CSPageProps {
   id: string
 }
 
 export default function CSPage({ id }: CSPageProps) {
-  const { data: problemDetail } = useCsProblemDetailQuery(Number(id))
-
-  // answered 필드로 답변 제출 여부 확인
-  const submitted = useMemo(() => {
-    if (!problemDetail) return false
-    return problemDetail.answered
-  }, [problemDetail])
+  const { submitted } = useCsPage({ id })
 
   return submitted ? <SubmittedDetailBox id={id} /> : <CsDetailBox id={id} />
 }
