+++ src/hooks/cs/useInfiniteScroll.ts
@@ -1,0 +1,28 @@
+'use client'
+
+import { useInView } from 'react-intersection-observer'
+import { useEffect } from 'react'
+
+interface UseInfiniteScrollProps {
+  fetchNextPage: () => void
+  hasNextPage: boolean | undefined
+  isFetchingNextPage: boolean
+}
+
+export const useInfiniteScroll = ({
+  fetchNextPage,
+  hasNextPage,
+  isFetchingNextPage,
+}: UseInfiniteScrollProps) => {
+  const { ref, inView } = useInView({
+    threshold: 0.1,
+  })
+
+  useEffect(() => {
+    if (inView && hasNextPage && !isFetchingNextPage) {
+      fetchNextPage()
+    }
+  }, [inView, hasNextPage, isFetchingNextPage, fetchNextPage])
+
+  return { ref }
+}
